Finish separating stream classes into separate files
