--- Streams/StreamWriter.cpp
@@ -1,3 +1,0 @@
-#include "StreamWriter.h"
-
-StreamWriter::~StreamWriter() { }
